# Conflicts:
#	CHANGELOG.md
#	docs/usage/extension.md
#	src/main/nextflow/co2footprint/CO2FootprintCLI.groovy
#	src/main/nextflow/co2footprint/CO2FootprintCalculator.groovy
#	src/main/nextflow/co2footprint/CO2FootprintExtension.groovy
#	src/main/nextflow/co2footprint/CO2FootprintFactory.groovy
#	src/main/nextflow/co2footprint/Config/BaseFileConfig.groovy
#	src/main/nextflow/co2footprint/FileCreation/ProvenanceFileCreator.groovy
#	src/main/nextflow/co2footprint/Records/CO2Record.groovy
#	src/main/nextflow/co2footprint/Records/CiRecordCollector.groovy
#	src/test/nextflow/co2footprint/CO2FootprintExtensionTest.groovy
#	src/test/nextflow/co2footprint/CO2FootprintPluginTest.groovy
#	src/test/nextflow/co2footprint/Records/CO2RecordTreeTest.groovy
#	src/testResources/cli/file_checks.json
#	src/testResources/cli/provenance_test.json
#	src/testResources/cli/report_test.html
#	src/testResources/cli/trace_test.txt
#	src/testResources/observer/file_checks.json
#	src/testResources/observer/report_test.html
#	src/testResources/report/test_payload.json
#	src/testResources/tree/file_checks.json
#	src/testResources/tree/test_tree.yaml

+++ src/main/nextflow/co2footprint/CO2FootprintCalculator.groovy
@@ -100,8 +100,8 @@ class CO2FootprintCalculator {
 
         /* ===== Memory Information ===== */
 
-        final Long requestedMemory = trace.get('memory') as Long        // [bytes]
-        final Long maxRequiredMemory = trace.get('peak_rss') as Long    // [bytes]
+        final BigInteger requestedMemory = trace.get('memory') as BigInteger        // [bytes]
+        final BigInteger maxRequiredMemory = trace.get('peak_rss') as BigInteger    // [bytes]
 
         // Assign the final memory value
         final BigDecimal memory
@@ -175,9 +175,9 @@ class CO2FootprintCalculator {
             ci,
             ciMarket,
             cpuUsage,
-            memory as Long,
+            memory,
             runtime_ms,
-            numberOfCores as Integer,
+            numberOfCores,
             pue,
             powerdrawPerCore,
             powerdrawMem,
